docs: :sparkles: add 404 page (#63)
## Description

This is a first try at a custom 404 page for our websites.

I'm pretty sure it doesn't work as we want it to yet. My questions:

- How do you link to the homepage of the website without writing it here
so it can be used across websites? I've added `/` but I don't think that
works.
- How do you link to an image on a 404 page? I played around with making
incorrect links and it seems that root of the 404 page changes depending
on where it is - e.g., `localhost/guide/incorrect-page` looks for an
image in the `guide` dir, while `localhost/incorrect-page` looks for an
image in the root dir?
- I'm not sure where the image should be located. Currently it's in
images bc that's there I would like it to be in each repo - unless
there's a way to link to it in a way so we don't need to have it in each
repo?

Screenshot: 

![image](https://github.com/user-attachments/assets/1e1f2d15-bd6d-4a89-a6f4-db51aa17d864)

Closes #55 

## Reviewer Focus

<!-- Please delete as appropriate: -->
This PR needs an in-depth review.

Focus on CHANGES.

## Checklist

- [X] Ran spell-check
- [X] Formatted Markdown
- [X] Rendered website locally (in the `seedcase-sprout` repo)

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>