style: :lipstick: add `brand.yml` to provide theming (#114)
# Description

This work very much mirrors what I’ve done in steno-aarhus/sdca-theme#11

The PR adds a `brand.yml` and integrates it in the `_extension.yml` and
`_quarto.yml`. I've made some minor changes/addition to the scss rules
and color palette that I will comment on at the specific lines in the
PR.

It also removes `theme.scss` since we don't need that anymore.

I've also added the `custom-styles.scss` and added it to `_quarto.yml`
to make it easier to layer themes, if we want different scss rules in
the different repos. Not sure that we want it for seedcase websites,
though. Thoughts?

Closes #106 
Closes #78 (maybe?)

This PR needs an in-depth review.

## Checklist

- [X] Ran `just run-all`