feat: :sparkles: add logos and font for `check-datapackage` (#164)
# Description

Let me know what you think about the logo. I was made rather quick. I
wanted it to fit with the other logos (shape and colour) and thought the
best "symbol" for checking would be a checkmark ✅

Once this has been merged (and synced to `check-datapackage`), I have
the first draft of the landing page ready.

Related to seedcase-project/check-datapackage#8.

(PS I don't care for the naming scheme of these files; we should
probably change it at some point).

This PR needs an in-depth review.